增加deploy/deploy-check-deps.sh: 检查软件包是否已安装 修复cluster-install中遇到的bug 输出脚本执行过程中过的日志到/tmp/log.txt 修复setup-ip.sh中过的master节点标记问题 修改deploy_env中的默认安装包路径到/root/dashboard_packages

+++ cluster/deploy/deploy-docker-deps.sh
@@ -2,6 +2,8 @@
 
 PACKAGE_PATH=${PACKAGE_PATH:-$HOME/dashboard_packages}
 DOCKER_VERSION=${DOCKER_VERSION:-"1.9.1"}
+
+source ./deploy-check-deps.sh
   
 function get_docker_deps_deb() {
   echo "Package path: $PACKAGE_PATH"
@@ -89,7 +91,7 @@ function get_docker_deps_deb() {
     fi
   fi
   echo " ... done"
-  
+ 
   echo -n "get libsystemd-journal0"
   if [ ! -f ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb ]; then
     ##libsystemd-journal0 204-5ubuntu20.19
@@ -122,10 +124,25 @@ function get_docker_deps_deb() {
 function install_docker_deps_apt() {
   ## Update package information, ensure that APT works with the https method, and that CA certificates are installed.
   sudo apt-get update
-  sudo apt-get -y install apt-transport-https ca-certificates
+  result=`check_deps apt-transport-https`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install apt-transport-https >& /dev/null
+  fi
+
+  result=`check_deps ca-certificates`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install ca-certificates >& /dev/null
+  fi
 
   ## For Ubuntu Trusty, Wily, and Xenial, it’s recommended to install the linux-image-extra-* kernel packages. The linux-image-extra-* packages allows you use the aufs storage driver.
-  sudo apt-get -y install linux-image-extra-$(uname -r) linux-image-extra-virtual
+  result=`check_deps linux-image-extra-$(uname -r)`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install linux-image-extra-$(uname -r) >& /dev/null
+  fi
+  result=`check_deps linux-image-extra-virtual`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install linux-image-extra-virtual >& /dev/null
+  fi
 
   ## Add the new GPG key.
   sudo apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
@@ -137,7 +154,10 @@ function install_docker_deps_apt() {
   sudo apt-get update
 
   ## Purge the old repo if it exists.
-  sudo apt-get purge lxc-docker
+  result=`check_deps lxc-docker`
+  if [ $result == "no" ]; then
+    sudo apt-get purge lxc-docker
+  fi
 
   ## Verify that APT is pulling from the right repository.
   sudo apt-cache policy docker-engine
@@ -149,30 +169,39 @@ function install_docker_deps_apt() {
   ##  aufs-tools cgroup-lite docker-engine libltdl7 libsystemd-journal0
 
   ## Install Docker.
-  sudo apt-get -y install docker-engine
+  result=`check_deps docker-engine`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install docker-engine >& /dev/null
+  fi
 }
 
 function install_docker_deps_dpkg() {
   ## ensure that APT works with the https method, and that CA certificates are installed."
-  echo -n "Install apt-transport-https"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/apt-transport-https_1.0.1ubuntu2.14_amd64.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - apt-transport-https_1.0.1ubuntu2.14_amd64.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  result=`check_deps apt-transport-https`
+  if [ $result == "no" ]; then
+    echo -n "Install apt-transport-https"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/apt-transport-https_1.0.1ubuntu2.14_amd64.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - apt-transport-https_1.0.1ubuntu2.14_amd64.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
   fi
-  echo " ... done"
 
-  echo -n "Install ca-certificates"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/ca-certificates_20160104ubuntu0.14.04.1_all.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - ca-certificates_20160104ubuntu0.14.04.1_all.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  result=`check_deps ca-certificates`
+  if [ $result == "no" ]; then
+    echo -n "Install ca-certificates"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/ca-certificates_20160104ubuntu0.14.04.1_all.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - ca-certificates_20160104ubuntu0.14.04.1_all.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
   fi
-  echo " ... done"
   
   ## "The linux-image-extra-* packages allows you use the aufs storage driver."
   ##TODO: specify a version or supply all versions
@@ -182,21 +211,43 @@ function install_docker_deps_dpkg() {
   echo "Purge the old repo if it exists."
   sudo apt-get purge lxc-docker >& /dev/null
   
-  echo "Install extra packages:  aufs-tools cgroup-lite libltdl7 libsystemd-journal0"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/aufs-tools_3.2+20130722-1.1_amd64.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/cgroup-lite_1.9_all.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/libltdl7_2.4.2-1.7ubuntu1_amd64.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb >& /dev/null
-  
-  echo -n "Install Docker"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  ## echo "Install extra packages:  aufs-tools cgroup-lite libltdl7 libsystemd-journal0"
+  result=`check_deps aufs-tools`
+  if [ $result == "no" ]; then
+    echo "Install extra packages:  aufs-tools"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/aufs-tools_3.2+20130722-1.1_amd64.deb >& /dev/null
   fi
-  echo " ... done"
+
+  result=`check_deps cgroup-lite`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: cgroup-lite"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/cgroup-lite_1.9_all.deb >& /dev/null
+  fi
+
+  result=`check_deps libltdl7`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: libltdl7"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/libltdl7_2.4.2-1.7ubuntu1_amd64.deb >& /dev/null
+  fi
+ 
+  result=`check_deps libsystemd-journal0`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: libsystemd-journal0"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb >& /dev/null
+  fi
+  
+  result=`check_deps docker-engine`
+  if [ $result == "no" ]; then
+    echo -n "Install Docker"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
+  fi
 }
 
 function uninstall_docker_deps() {
